+++ assignment-2/arduino/src/iot/sensor/sonar/Sonar.h
@@ -2,17 +2,18 @@
 #define _SONAR_H_
 
 #include "../Sensor.h"
+#include <Arduino.h>
 
 class Sonar : public Sensor {
 
 private:
     const float vs = 331.45 + 0.62 * 20;
-    int trigPin;
-    int echoPin;
+    uint8_t trigPin;
+    uint8_t echoPin;
     float distance;
 
 public:
-    Sonar(int trigPin, int echoPin);
+    Sonar(uint8_t trigPin, uint8_t echoPin);
     float getDistanceInCM();
 
     void setup();
