+++ assignment-2/arduino/src/iot/sensor/thermometer/Thermometer.cpp
@@ -1,6 +1,6 @@
 #include "Thermometer.h"
 
-Thermometer::Thermometer(int pin) {
+Thermometer::Thermometer(uint8_t pin) {
     this->pin = pin;
 }
