+++ assignment-2/arduino/src/smart-coffee-machine/sensors/sugar/Sugar.cpp
@@ -1,6 +1,6 @@
 #include "Sugar.h"
 
-Sugar::Sugar(int pin) : Potentiometer(pin) {}
+Sugar::Sugar(uint8_t pin) : Potentiometer(pin) {}
 
 int Sugar::getPercentage() {
     return map(this->getValue(), 0, 1023, 0, 100);
