+++ assignment-2/arduino/src/smart-coffee-machine/sensors/sugar/Sugar.h
@@ -6,7 +6,7 @@
 class Sugar : public Potentiometer {
 
 public:
-    Sugar(int pin);
+    Sugar(uint8_t pin);
     int getPercentage();
 
 };
